refactor: 모듈 구조 개선 및 Import 방식 정리
- 패키지 초기화 개선: crud, schemas, models 패키지의 __init__.py를 수정하여 내부 모듈을 자동으로 노출하도록 변경 (* import 사용)
- 의존성 파일명 변경 및 정의 파편화 통합: app/api/dependency.py -> app/api/deps.py
- Import 정리: 모듈 자체를 import 하는 방식(예: from app.crud import user)을 제거하고, 함수/클래스를 직접 import 하도록 통일하여 네임스페이스 충돌 방지
- 코드 정리: 변경된 구조에 맞춰 API 라우터 및 테스트 코드의 import 구문 일괄 수정
- 파일 분리 과정에서 생기는 순환참조 오류 해결
