+++ app/alembic/env.py
@@ -3,8 +3,8 @@ from logging.config import fileConfig
 from alembic import context
 from sqlalchemy import engine_from_config, pool
 
-from app.config import settings
-from app.models import BaseModel
+from app.core.config import settings
+from app.models.models import BaseModel
 
 # this is the Alembic Config object, which provides
 # access to the values within the .ini file in use.
