--- app/api/dependency.py
@@ -1,27 +1,0 @@
-from typing import Annotated
-
-from fastapi import Depends, HTTPException, status
-from fastapi.security import OAuth2PasswordBearer
-
-from app.core.security import verify_access_token
-from app.database import SessionDep  # noqa: TCH001
-from app.models import User
-
-oauth2_scheme = OAuth2PasswordBearer(tokenUrl="/v1/auth/login")
-TokenDep = Annotated[str, Depends(oauth2_scheme)]
-
-
-async def get_current_user(*, session: SessionDep, token: TokenDep):
-    data = verify_access_token(token)
-    if not data:
-        raise HTTPException(
-            status_code=status.HTTP_401_UNAUTHORIZED,
-            detail="Not authenticated",
-            headers={"WWW-Authenticate": "Bearer"},
-        )
-
-    user = session.get(User, data["sub"])
-    return user
-
-
-AuthUserDep = Annotated[User, Depends(get_current_user)]
