+++ app/models/__init__.py
@@ -1,0 +1,3 @@
+from .models import Post, PostComment, User
+
+__all__ = ["Post", "PostComment", "User"]
