+++ tests/test_user.py
@@ -4,8 +4,8 @@ from fastapi import status
 from sqlalchemy import select
 from sqlalchemy.orm import Session
 
-from app import query
 from app.core.security import verify_password
+from app.crud import get_user_by_email
 from app.models import User
 from tests.utils import DEFAULT_USER_EMAIL, DEFAULT_USER_NICKNAME, random_email, random_lower_string
 
@@ -67,7 +67,7 @@ def test_update_user_me(session: Session, client: TestClient, default_user_token
     result = client.patch("/v1/user/me", headers=default_user_token_header, json=data)
     data = result.json()
 
-    user = query.get_user_by_email(session=session, email=DEFAULT_USER_EMAIL)
+    user = get_user_by_email(session=session, email=DEFAULT_USER_EMAIL)
 
     assert result.status_code == status.HTTP_200_OK
     assert user.nickname == data["nickname"]
@@ -78,6 +78,6 @@ def test_update_user_me_without_data(
 ) -> None:
     data = {}
     result = client.patch("/v1/user/me", headers=default_user_token_header, json=data)
-    user = query.get_user_by_email(session=session, email=DEFAULT_USER_EMAIL)
+    user = get_user_by_email(session=session, email=DEFAULT_USER_EMAIL)
     assert result.status_code == status.HTTP_200_OK
     assert user.nickname == DEFAULT_USER_NICKNAME
